Renamed the modify the image size sample
