--- Markdown-to-Word-conversion/Modify_image_size/Modify_image_size/Data/Input.md
@@ -1,3 +1,0 @@
-Image from file:
-
-![File](Image_2.png)
